Generalize runConcurrential to use arbitrary monad
I have attempted to define transformers ExceptionSafeT and SharedStateT
to be used over top of Concurrential, but had a lot of trouble combining
them, since the ExceptionSafeT needs access to the raw IO, raw IO which
SharedStateT instances cannot provide.

So, I've taken a step back, and attempted to build extensibility into
runConcurrential(K)

This is very bizarre and as far as I know unprecedented. Let's see how
it goes!