Fix "blind bind" problem via MVar
The issue: Concurrential was blind to sequentiality through binds. A
sequential term to the right of a bind which contains sequential terms
will run  concurrently with those sequential terms! That's because we
did not pass on the sequential part of the bind's continuation.

So, before this change, the term

     (concurrently a >> sequentially b)
  *> sequentially c

would not guarantee b before c ! Now it does.